https://github.com/sathishpy/corrugation/issues/8 https://github.com/sathishpy/corrugation/issues/5 Add the required data types for Paper Roll report so thar they cn be filtered properly
Add provison for data importing - mainly customer, supplier, paper roll and box. Ideally
we should be using the ERPNext built in import tools. However it makes more sense to
provide an abstraction to hide the details. This is just a basic draft, so certainly needs
to improved along the lines of offical data import tool